% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_words.R
\name{count_words}
\alias{count_words}
\title{count_words}
\usage{
count_words(
  text_data,
  include_2gram = F,
  stopwords = load_stopwords(),
  mention_limit = 10,
  patterns = NULL
)
}
\arguments{
\item{text_data}{data.frame with text as first column}

\item{include_2gram}{boolean}

\item{stopwords}{list of stopwords}

\item{mention_limit}{int of where to cutoff words}

\item{patterns}{list of named replacements of the form "regex" = "replacement"}
}
\value{
a dataframe with words and their counts
}
\description{
count_words
}
