% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cloud.R
\name{create_cloud}
\alias{create_cloud}
\title{create_cloud}
\usage{
create_cloud(text_col, path, ...)
}
\arguments{
\item{text_col}{character}

\item{path}{character - path to data}

\item{...}{fundals passed to count_words.}
}
\value{
wordcloud figure
}
\description{
create_cloud
}
