% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{load_data}
\usage{
load_data(text_col, path)
}
\arguments{
\item{text_col}{character}

\item{path}{character - path to the raw data}
}
\value{
text for count_words
}
\description{
load_data
}
